refactored near2far
