+++ tests/test_components/test_IO.py
@@ -15,10 +15,10 @@ from tidy3d import *
 from tidy3d import __version__
 import tidy3d as td
 from tidy3d.components.base import Tidy3dBaseModel
-from .utils import SIM_FULL as SIM
-from .utils import SIM_MONITORS as SIM2
-from .utils import clear_tmp
-from .test_data_monitor import make_flux_data
+from ..utils import SIM_FULL as SIM
+from ..utils import SIM_MONITORS as SIM2
+from ..utils import clear_tmp
+from ..test_data.test_monitor_data import make_flux_data
 
 # Store an example of every minor release simulation to test updater in the future
 SIM_DIR = "tests/sims"
