+++ tests/test_data/test_monitor_data.py
@@ -22,7 +22,7 @@ from .test_dataset import make_flux_data, make_flux_time_data
 from .test_data_arrays import FIELD_MONITOR, FIELD_TIME_MONITOR, MODE_SOLVE_MONITOR
 from .test_data_arrays import MODE_MONITOR, PERMITTIVITY_MONITOR, FLUX_MONITOR, FLUX_TIME_MONITOR
 from .test_data_arrays import SIM_SYM
-from .utils import clear_tmp
+from ..utils import clear_tmp
 
 """ Make the montor data """
