+++ tests/test_package/test_config.py
@@ -6,7 +6,7 @@ import tidy3d as td
 from tidy3d.web.config import DEFAULT_CONFIG, WEB_CONFIGS
 from tidy3d.log import log, DEFAULT_LEVEL, LEVEL_MAP
 
-from .utils import assert_log_level
+from ..utils import assert_log_level
 
 
 def test_logging_level():
